Cree el CRUD de metodo_pago (el delete es logico)

+++ src/main/java/mx/tecnm/backend/api/models/Metodo_pago.java
@@ -1,5 +1,5 @@
 package mx.tecnm.backend.api.models;
 
-public record Metodo_pago (int id, String nombre, int comision, String estado){
+public record Metodo_pago (int id, String nombre, double comision, String estado){
 
 }
